npm audit MongoMemoryServer version update

+++ tests/db-handler.js
@@ -3,13 +3,15 @@
 const mongoose = require('mongoose');
 const { MongoMemoryServer } = require('mongodb-memory-server');
 
-const mongod = new MongoMemoryServer();
+let mongod = undefined;
+
 
 /**
  * Connect to the in-memory database.
  */
 module.exports.connect = async () => {
-    const uri = await mongod.getConnectionString();
+    mongod = await MongoMemoryServer.create();
+    const uri = mongod.getUri();
 
     const mongooseOpts = {
         useNewUrlParser: true,
@@ -25,19 +27,23 @@ module.exports.connect = async () => {
  * Drop database, close the connection and stop mongod.
  */
 module.exports.closeDatabase = async () => {
-    await mongoose.connection.dropDatabase();
-    await mongoose.connection.close();
-    await mongod.stop();
+    if (mongod) {
+        await mongoose.connection.dropDatabase();
+        await mongoose.connection.close();
+        await mongod.stop();
+    }
 };
 
 /**
  * Remove all the data for all db collections.
  */
 module.exports.clearDatabase = async () => {
-    const collections = mongoose.connection.collections;
+    if (mongod) {
+        const collections = mongoose.connection.collections;
 
-    for (const key in collections) {
-        const collection = collections[key];
-        await collection.deleteMany();
+        for (const key in collections) {
+            const collection = collections[key];
+            await collection.deleteMany();
+        }
     }
 };
